@@ -1,9 +1,9 @@
 {
-  "name": "notus-angular",
+  "name": "1.0.0",
   "version": "1.1.0",
-  "description": "Notus Angular - Free Tailwind CSS UI Kit and Admin by Creative Tim.",
-  "repository": "https://github.com/creativetimofficial/notus-angular",
-  "license": "MIT",
+  "description": "Livmoney",
+  "repository": "",
+  "license": "LIV",
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
@@ -15,29 +15,34 @@
     "install:clean": "rm -rf node_modules/ && rm -rf package-lock.json && npm install && npm run build:tailwind && npm start"
   },
   "dependencies": {
-    "@angular/animations": "11.2.5",
-    "@angular/common": "11.2.5",
-    "@angular/compiler": "11.2.5",
-    "@angular/core": "11.2.5",
-    "@angular/forms": "11.2.5",
-    "@angular/platform-browser": "11.2.5",
-    "@angular/platform-browser-dynamic": "11.2.5",
-    "@angular/router": "11.2.5",
+    "@angular/animations": "^12.2.2",
+    "@angular/cdk": "^12.2.2",
+    "@angular/common": "^12.2.2",
+    "@angular/compiler": "12.2.2",
+    "@angular/core": "^12.2.2",
+    "@angular/forms": "^12.2.2",
+    "@angular/platform-browser": "^12.2.2",
+    "@angular/platform-browser-dynamic": "^12.2.2",
+    "@angular/router": "^12.2.2",
     "@fortawesome/fontawesome-free": "5.15.3",
     "@popperjs/core": "2.9.1",
     "@tailwindcss/forms": "0.2.1",
+    "angularx-social-login": "^4.0.1",
     "chart.js": "2.9.4",
     "ngx-toastr": "^14.1.0",
+    "primeng": "^11.4.5",
     "rxjs": "6.6.6",
+    "rxjs-compat": "^6.6.7",
+    "sweetalert2": "^11.1.4",
     "tailwindcss": "2.0.4",
-    "tslib": "2.1.0",
+    "tslib": "^2.3.1",
     "zone.js": "0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "0.1102.4",
-    "@angular/cli": "11.2.4",
-    "@angular/compiler-cli": "11.2.5",
-    "@angular/language-service": "11.2.5",
+    "@angular-devkit/build-angular": "^0.1102.14",
+    "@angular/cli": "^12.2.2",
+    "@angular/compiler-cli": "12.2.2",
+    "@angular/language-service": "12.2.2",
     "@types/jasmine": "3.6.6",
     "@types/jasminewd2": "2.0.8",
     "@types/node": "14.14.35",
@@ -52,6 +57,6 @@
     "protractor": "7.0.0",
     "ts-node": "9.1.1",
     "tslint": "6.1.3",
-    "typescript": "4.1.5"
+    "typescript": "^4.3.5"
   }
 }
